+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added the `craft\commerce\stripe\gateways\Gateway::EVENT_BEFORE_SUBSCRIBE` event that is fired before subscribing a customer to a subscription plan. ([#31](https://github.com/craftcms/commerce-stripe/issues/31))
 
 ### Changed
+- Switch Stripe API version to '2018-09-24'.
 - Stripe gateway now throws an exception when failing to process a webhook.
 - Stripe plugin now requires Craft Commerce `2.0.0-beta.12`.